Moved ParserUtils into Parser and Parser into Query
