+++ src/Expression/Parse.hs
@@ -5,7 +5,7 @@ module Expression.Parse
   ) where
 
 import Expression.Expr (Ex(..))
-import ParseUtils (Parser, lex0)
+import Parser (Parser, lex0)
 import Text.Megaparsec
 import Text.Megaparsec.Char
 import qualified Text.Megaparsec.Char.Lexer as L
