+++ src/Table/Utils.hs
@@ -1,4 +1,4 @@
-module Utils
+module Table.Utils
   ( omap
   , okeys
   , ovals
@@ -7,7 +7,7 @@ module Utils
 import Data.Map.Ordered (OMap, assocs, fromList)
 
 omap :: Ord a => (b -> c) -> OMap a b -> OMap a c
-omap f om = fromList $ fmap (\(k, v) -> (k, f v)) (assocs om)
+omap f = fromList . fmap (\(k, v) -> (k, f v)) . assocs
 
 okeys :: OMap a b -> [a]
 okeys = fmap fst . assocs
