+++ test/fixtures/golden/select_star_as_fail.csv
@@ -1,4 +1,4 @@
-1:10:
+SQL Query Parser:1:10:
   |
 1 | SELECT * AS star
   |          ^^^^
