+++ test/fixtures/golden/select_type_fail_no_as.csv
@@ -1,4 +1,4 @@
-1:10:
+SQL Query Parser:1:10:
   |
 1 | SELECT 2 + 3
   |          ^^^^
